Complete class validation and publication stages
